Refactor Google API sync functions to use shared fetch helper
- Replaced individual fetch implementations in Gmail, Google Drive, and Google Photos connectors with a new shared `fetchGoogleJson` function.
- This change centralizes error handling and rate limiting for Google API requests, improving code maintainability and reducing duplication.
- Updated relevant imports and adjusted function signatures accordingly.

This refactor enhances the robustness of the Google API integrations and streamlines the codebase.
