+++ packages/gateway/src/connectors/google-sync-shared.ts
@@ -1,3 +1,5 @@
+import type { SyncContext } from "../sync/types.ts";
+import { UnauthenticatedError } from "../sync/types.ts";
 import { asUnknownObjectRecord } from "./json-unknown.ts";
 
 /** Best-effort summary from Google API JSON error bodies (never log raw tokens). */
@@ -32,4 +34,42 @@ export function formatGoogleHttpError(status: number, bodyText: string, service:
     return `${base} — ${oneLine.length > max ? `${oneLine.slice(0, max)}…` : oneLine}`;
   }
   return base;
+}
+
+/**
+ * Shared fetch helper for all Google API connectors.
+ * Acquires the "google" rate-limit token, merges the Bearer header, and parses JSON.
+ * Throws UnauthenticatedError on 401; uses formatGoogleHttpError for all other non-2xx responses.
+ */
+export async function fetchGoogleJson(
+  ctx: SyncContext,
+  token: string,
+  url: string,
+  service: string,
+  init?: RequestInit,
+): Promise<{ json: unknown; bytes: number }> {
+  await ctx.rateLimiter.acquire("google");
+  const merged = new Headers({ Authorization: `Bearer ${token}` });
+  if (init?.headers !== undefined) {
+    for (const [k, v] of new Headers(init.headers)) {
+      merged.set(k, v);
+    }
+  }
+  const res = await fetch(url, { ...init, headers: merged });
+  const text = await res.text();
+  const bytes = Buffer.byteLength(text, "utf8");
+  if (!res.ok) {
+    const msg = formatGoogleHttpError(res.status, text, service);
+    if (res.status === 401) {
+      throw new UnauthenticatedError(msg);
+    }
+    throw new Error(msg);
+  }
+  let json: unknown;
+  try {
+    json = JSON.parse(text) as unknown;
+  } catch {
+    throw new Error(`${service} sync failed: invalid JSON`);
+  }
+  return { json, bytes };
 }
