+++ packages/gateway/src/ipc/server.ts
@@ -20,18 +20,14 @@ import { ConnectorRpcError, dispatchConnectorRpc } from "./connector-rpc.ts";
 import { ConsentCoordinatorImpl } from "./consent.ts";
 import { DiagnosticsRpcError, dispatchDiagnosticsRpc } from "./diagnostics-rpc.ts";
 import {
-  encodeLine,
   errorResponse,
   isRequest,
   type JsonRpcId,
   type JsonRpcNotification,
-  type JsonRpcOutbound,
-  JsonRpcParseError,
   type JsonRpcRequest,
-  NdjsonLineReader,
-  parseJsonRpcLine,
 } from "./jsonrpc.ts";
 import { dispatchPeopleRpc, PeopleRpcError } from "./people-rpc.ts";
+import { ClientSession, type SessionWrite } from "./session.ts";
 import { dispatchSessionRpc, SessionRpcError } from "./session-rpc.ts";
 import type { IPCServer } from "./types.ts";
 import type { WorkflowRunContext, WorkflowRunHandler } from "./workflow-invoke.ts";
@@ -119,117 +115,7 @@ async function dispatchVaultIfPresent(
   }
 }
 
-type SessionWrite = (line: string) => void;
-
 type BunSessionData = { session: ClientSession };
-
-class ClientSession {
-  readonly clientId: string;
-  private readonly reader = new NdjsonLineReader();
-  private readonly write: SessionWrite;
-  private readonly onRpc: (
-    clientId: string,
-    msg: JsonRpcRequest | JsonRpcNotification,
-  ) => void | Promise<void>;
-  private readonly onDispose: (clientId: string) => void;
-  private disposed = false;
-
-  constructor(
-    clientId: string,
-    write: SessionWrite,
-    onRpc: (clientId: string, msg: JsonRpcRequest | JsonRpcNotification) => void | Promise<void>,
-    onDispose: (clientId: string) => void,
-  ) {
-    this.clientId = clientId;
-    this.write = write;
-    this.onRpc = onRpc;
-    this.onDispose = onDispose;
-  }
-
-  push(chunk: Uint8Array): void {
-    if (this.disposed) {
-      return;
-    }
-    let lines: string[];
-    try {
-      lines = this.reader.push(chunk);
-    } catch (e) {
-      this.sendParseFailure(e);
-      return;
-    }
-    void this.dispatchLines(lines).catch((e: unknown) => {
-      const m = e instanceof Error ? e.message : "dispatch error";
-      this.write(encodeLine(errorResponse(null, -32603, m)));
-      this.dispose();
-    });
-  }
-
-  endInput(): void {
-    if (this.disposed) {
-      return;
-    }
-    let lines: string[];
-    try {
-      lines = this.reader.flush();
-    } catch (e) {
-      this.sendParseFailure(e);
-      return;
-    }
-    void this.dispatchLines(lines).catch((e: unknown) => {
-      const m = e instanceof Error ? e.message : "dispatch error";
-      this.write(encodeLine(errorResponse(null, -32603, m)));
-      this.dispose();
-    });
-  }
-
-  writeOutbound(msg: JsonRpcOutbound): void {
-    if (this.disposed) {
-      return;
-    }
-    this.write(encodeLine(msg));
-  }
-
-  writeNotification(n: JsonRpcNotification): void {
-    this.writeOutbound(n);
-  }
-
-  private sendParseFailure(e: unknown): void {
-    const msg = e instanceof JsonRpcParseError ? e.message : "Parse error";
-    this.write(encodeLine(errorResponse(null, -32700, msg)));
-    this.dispose();
-  }
-
-  private async dispatchLines(lines: string[]): Promise<void> {
-    for (const line of lines) {
-      let msg: JsonRpcRequest | JsonRpcNotification;
-      try {
-        msg = parseJsonRpcLine(line);
-      } catch (e) {
-        const m = e instanceof JsonRpcParseError ? e.message : "Parse error";
-        this.writeOutbound(errorResponse(null, -32700, m));
-        continue;
-      }
-      try {
-        await this.onRpc(this.clientId, msg);
-      } catch (e) {
-        try {
-          const m = e instanceof Error ? e.message : "Internal error";
-          this.writeOutbound(errorResponse(null, -32603, m));
-        } catch {
-          this.dispose();
-        }
-      }
-    }
-  }
-
-  dispose(): void {
-    if (this.disposed) {
-      return;
-    }
-    this.disposed = true;
-    this.onDispose(this.clientId);
-  }
-}
 
 export type CreateIpcServerOptions = {
   listenPath: string;
